+++ cardinal/ecs/world_context.go
@@ -6,7 +6,7 @@ import (
 	"github.com/rs/zerolog"
 	ecslog "pkg.world.dev/world-engine/cardinal/ecs/log"
 	"pkg.world.dev/world-engine/cardinal/ecs/store"
-	"pkg.world.dev/world-engine/cardinal/txqueue"
+	"pkg.world.dev/world-engine/cardinal/txpool"
 )
 
 type WorldContext interface {
@@ -18,7 +18,7 @@ type WorldContext interface {
 	GetWorld() *World
 	StoreReader() store.Reader
 	StoreManager() store.IManager
-	GetTxQueue() *txqueue.TxQueue
+	GetTxQueue() *txpool.TxQueue
 	IsReadOnly() bool
 }
 
@@ -28,12 +28,12 @@ var (
 
 type worldContext struct {
 	world    *World
-	txQueue  *txqueue.TxQueue
+	txQueue  *txpool.TxQueue
 	logger   *ecslog.Logger
 	readOnly bool
 }
 
-func NewWorldContextForTick(world *World, queue *txqueue.TxQueue, logger *ecslog.Logger) WorldContext {
+func NewWorldContextForTick(world *World, queue *txpool.TxQueue, logger *ecslog.Logger) WorldContext {
 	return &worldContext{
 		world:    world,
 		txQueue:  queue,
@@ -72,7 +72,7 @@ func (w *worldContext) GetWorld() *World {
 	return w.world
 }
 
-func (w *worldContext) GetTxQueue() *txqueue.TxQueue {
+func (w *worldContext) GetTxQueue() *txpool.TxQueue {
 	return w.txQueue
 }
