+++ cardinal/types/message/message_test.go
@@ -7,9 +7,8 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"pkg.world.dev/world-engine/cardinal/txqueue"
-
 	"pkg.world.dev/world-engine/assert"
+	"pkg.world.dev/world-engine/cardinal/txpool"
 
 	"pkg.world.dev/world-engine/cardinal/ecs"
 	"pkg.world.dev/world-engine/cardinal/testutils"
@@ -591,7 +590,7 @@ func TestCopyTransactions(t *testing.T) {
 	type FooMsg struct {
 		X int
 	}
-	txq := txqueue.NewTxQueue()
+	txq := txpool.NewTxQueue()
 	txq.AddTransaction(1, FooMsg{X: 3}, &sign.Transaction{PersonaTag: "foo"})
 	txq.AddTransaction(2, FooMsg{X: 4}, &sign.Transaction{PersonaTag: "bar"})
